Save the accepted MCMC parameter values as floats
The MCMC parameter values don't have numeric error build up, and have
significantly more statistical uncertainty than the numeric error
introduced by truncating the values.  A lot of output file space is saved
by dropping the insignificant bits.

+++ src/Fitter/include/MCMCEngine.h
@@ -33,7 +33,7 @@ protected:
   void FillPoints();
 
   // The point value that is associated with the last call to the likelihood.
-  std::vector<double> fPoint;
+  std::vector<float> fPoint;
 };
 #endif
